requirements.txt
google-cloud-videointelligence==2.6.1
google-cloud-storage==2.0.0
moviepy==1.0.3